requirements.txt
fastapi>=0.111,<0.116
uvicorn[standard]>=0.30,<0.33
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<1.7
joblib>=1.3,<2.0
lightgbm>=4.3,<5.0
yfinance>=0.2.40,<0.3
feedparser>=6.0,<7.0
httpx>=0.27,<0.29
apscheduler>=3.10,<4.0
alpaca-py>=0.30,<1.0
supabase>=2.4,<3.0
transformers>=4.41,<5.0
torch>=2.3,<2.6
lightning>=2.3,<2.5
pytorch-forecasting>=1.0,<1.2
neuralforecast>=1.7,<2.0
stable-baselines3>=2.3,<3.0
gymnasium>=0.29,<1.1
praw>=7.7,<8.0
